# service/service/employee/employee-core/src/main/java/bs/service/employee/controller/generated/CourseSessionController.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package bs.service.employee.controller.generated;

import bs.lib.common.model.generated.NewRecordVTO;
import bs.lib.common.model.vto.ErrorVTO;
import bs.lib.sql.db.adapter.model.generated.OrderDirections;
import bs.service.employee.model.enums.SessionStatus;
import bs.service.employee.model.generated.CourseSessionDTO;
import bs.service.employee.model.generated.CourseSessionResultSet;
import bs.service.employee.model.generated.CourseSessionVTO;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.time.LocalDate;
import org.springframework.format.annotation.DateTimeFormat;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.annotation.Secured;
import org.springframework.stereotype.Controller;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.6.0")
@Validated
@Controller
@Tag(name = "CourseSession", description = "the CourseSession API")
public interface CourseSessionController {

    /**
     * POST /courses/{courseId}/sessions : Create course session
     *
     * @param courseId
     *            (required)
     * @param courseSessionDTO
     *            (required)
     *
     * @return OK (status code 200) or Bad Request (status code 400)
     */
    @Operation(operationId = "createCourseSession", summary = "Create course session", tags = {
            "CourseSession" }, responses = { @ApiResponse(responseCode = "200", description = "OK", content = {
                    @Content(mediaType = "application/json", schema = @Schema(implementation = NewRecordVTO.class)) }),
                    @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class)) }) })
    @RequestMapping(method = RequestMethod.POST, value = "/courses/{courseId}/sessions", produces = {
            "application/json" }, consumes = { "application/json" })

    ResponseEntity<NewRecordVTO> _createCourseSession(
            @Parameter(name = "courseId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("courseId") Integer courseId,
            @Parameter(name = "CourseSessionDTO", description = "", required = true) @Valid @RequestBody CourseSessionDTO courseSessionDTO);

    /**
     * DELETE /courses/{courseId}/sessions/{sessionId} : Delete course session (soft delete)
     *
     * @param courseId
     *            (required)
     * @param sessionId
     *            (required)
     *
     * @return OK (status code 200) or Bad Request (status code 400)
     */
    @Operation(operationId = "deleteCourseSession", summary = "Delete course session (soft delete)", tags = {
            "CourseSession" }, responses = { @ApiResponse(responseCode = "200", description = "OK"),
                    @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class)) }) })
    @RequestMapping(method = RequestMethod.DELETE, value = "/courses/{courseId}/sessions/{sessionId}", produces = {
            "application/json" })

    ResponseEntity<Void> _deleteCourseSession(
            @Parameter(name = "courseId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("courseId") Integer courseId,
            @Parameter(name = "sessionId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("sessionId") Integer sessionId);

    /**
     * GET /courses/{courseId}/sessions : Get all course sessions by filter
     *
     * @param courseId
     *            (required)
     * @param status
     *            (optional)
     * @param sessionDateFrom
     *            (optional)
     * @param sessionDateTo
     *            (optional)
     * @param startTimeFrom
     *            (optional)
     * @param startTimeTo
     *            (optional)
     * @param endTimeFrom
     *            (optional)
     * @param endTimeTo
     *            (optional)
     * @param pageNum
     *            (optional, default to 0)
     * @param pageSize
     *            (optional, default to 25)
     * @param orderDir
     *            Order Direction (optional)
     * @param orderBy
     *            Order By Attribute (optional)
     *
     * @return OK (status code 200) or Bad Request (status code 400)
     */
    @Operation(operationId = "getAllCourseSessionsByFilter", summary = "Get all course sessions by filter", tags = {
            "CourseSession" }, responses = { @ApiResponse(responseCode = "200", description = "OK", content = {
                    @Content(mediaType = "application/json", schema = @Schema(implementation = CourseSessionResultSet.class)) }),
                    @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class)) }) })
    @RequestMapping(method = RequestMethod.GET, value = "/courses/{courseId}/sessions", produces = {
            "application/json" })

    ResponseEntity<CourseSessionResultSet> _getAllCourseSessionsByFilter(
            @Parameter(name = "courseId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("courseId") Integer courseId,
            @Parameter(name = "status", description = "", in = ParameterIn.QUERY) @Valid SessionStatus status,
            @Parameter(name = "sessionDateFrom", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sessionDateFrom", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate sessionDateFrom,
            @Parameter(name = "sessionDateTo", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sessionDateTo", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate sessionDateTo,
            @Parameter(name = "startTimeFrom", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "startTimeFrom", required = false) String startTimeFrom,
            @Parameter(name = "startTimeTo", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "startTimeTo", required = false) String startTimeTo,
            @Parameter(name = "endTimeFrom", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "endTimeFrom", required = false) String endTimeFrom,
            @Parameter(name = "endTimeTo", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "endTimeTo", required = false) String endTimeTo,
            @Min(0) @Parameter(name = "pageNum", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageNum", required = false, defaultValue = "0") Integer pageNum,
            @Min(1) @Max(100) @Parameter(name = "pageSize", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageSize", required = false, defaultValue = "25") Integer pageSize,
            @Parameter(name = "orderDir", description = "Order Direction", in = ParameterIn.QUERY) @Valid @RequestParam(value = "orderDir", required = false) OrderDirections orderDir,
            @Parameter(name = "orderBy", description = "Order By Attribute", in = ParameterIn.QUERY) @Valid @RequestParam(value = "orderBy", required = false) String orderBy);

    /**
     * GET /courses/sessions : Get all course sessions across all courses with filters
     *
     * @param courseId
     *            (optional)
     * @param trainerId
     *            (optional)
     * @param placeId
     *            (optional)
     * @param status
     *            (optional)
     * @param sessionDateFrom
     *            (optional)
     * @param sessionDateTo
     *            (optional)
     * @param startTimeFrom
     *            (optional)
     * @param startTimeTo
     *            (optional)
     * @param endTimeFrom
     *            (optional)
     * @param endTimeTo
     *            (optional)
     * @param pageNum
     *            (optional, default to 0)
     * @param pageSize
     *            (optional, default to 25)
     * @param orderDir
     *            Order Direction (optional)
     * @param orderBy
     *            Order By Attribute (optional)
     *
     * @return OK (status code 200) or Bad Request (status code 400)
     */
    @Operation(operationId = "getAllSessionsByFilter", summary = "Get all course sessions across all courses with filters", tags = {
            "CourseSession" }, responses = { @ApiResponse(responseCode = "200", description = "OK", content = {
                    @Content(mediaType = "application/json", schema = @Schema(implementation = CourseSessionResultSet.class)) }),
                    @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class)) }) })
    @RequestMapping(method = RequestMethod.GET, value = "/courses/sessions", produces = { "application/json" })

    ResponseEntity<CourseSessionResultSet> _getAllSessionsByFilter(
            @Parameter(name = "courseId", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "courseId", required = false) Integer courseId,
            @Parameter(name = "trainerId", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "trainerId", required = false) Integer trainerId,
            @Parameter(name = "placeId", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "placeId", required = false) Integer placeId,
            @Parameter(name = "status", description = "", in = ParameterIn.QUERY) @Valid SessionStatus status,
            @Parameter(name = "sessionDateFrom", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sessionDateFrom", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate sessionDateFrom,
            @Parameter(name = "sessionDateTo", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sessionDateTo", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate sessionDateTo,
            @Parameter(name = "startTimeFrom", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "startTimeFrom", required = false) String startTimeFrom,
            @Parameter(name = "startTimeTo", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "startTimeTo", required = false) String startTimeTo,
            @Parameter(name = "endTimeFrom", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "endTimeFrom", required = false) String endTimeFrom,
            @Parameter(name = "endTimeTo", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "endTimeTo", required = false) String endTimeTo,
            @Min(0) @Parameter(name = "pageNum", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageNum", required = false, defaultValue = "0") Integer pageNum,
            @Min(1) @Max(100) @Parameter(name = "pageSize", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageSize", required = false, defaultValue = "25") Integer pageSize,
            @Parameter(name = "orderDir", description = "Order Direction", in = ParameterIn.QUERY) @Valid @RequestParam(value = "orderDir", required = false) OrderDirections orderDir,
            @Parameter(name = "orderBy", description = "Order By Attribute", in = ParameterIn.QUERY) @Valid @RequestParam(value = "orderBy", required = false) String orderBy);

    /**
     * GET /courses/{courseId}/sessions/{sessionId} : Get course session by id
     *
     * @param courseId
     *            (required)
     * @param sessionId
     *            (required)
     *
     * @return OK (status code 200) or Bad Request (status code 400)
     */
    @Operation(operationId = "getCourseSessionById", summary = "Get course session by id", tags = {
            "CourseSession" }, responses = { @ApiResponse(responseCode = "200", description = "OK", content = {
                    @Content(mediaType = "application/json", schema = @Schema(implementation = CourseSessionVTO.class)) }),
                    @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class)) }) })
    @RequestMapping(method = RequestMethod.GET, value = "/courses/{courseId}/sessions/{sessionId}", produces = {
            "application/json" })

    ResponseEntity<CourseSessionVTO> _getCourseSessionById(
            @Parameter(name = "courseId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("courseId") Integer courseId,
            @Parameter(name = "sessionId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("sessionId") Integer sessionId);

    /**
     * PUT /courses/{courseId}/sessions/{sessionId} : Update course session
     *
     * @param courseId
     *            (required)
     * @param sessionId
     *            (required)
     * @param courseSessionDTO
     *            (required)
     *
     * @return OK (status code 200) or Bad Request (status code 400)
     */
    @Operation(operationId = "updateCourseSession", summary = "Update course session", tags = {
            "CourseSession" }, responses = { @ApiResponse(responseCode = "200", description = "OK", content = {
                    @Content(mediaType = "application/json", schema = @Schema(implementation = NewRecordVTO.class)) }),
                    @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorVTO.class)) }) })
    @RequestMapping(method = RequestMethod.PUT, value = "/courses/{courseId}/sessions/{sessionId}", produces = {
            "application/json" }, consumes = { "application/json" })

    ResponseEntity<NewRecordVTO> _updateCourseSession(
            @Parameter(name = "courseId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("courseId") Integer courseId,
            @Parameter(name = "sessionId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("sessionId") Integer sessionId,
            @Parameter(name = "CourseSessionDTO", description = "", required = true) @Valid @RequestBody CourseSessionDTO courseSessionDTO);

}